# Improve performance of interleave_primitive (-15% - 45%) / interleave_bytes (-10-25%)