@Dandandan 2025-04-18
> Awesome! Makes me want to look at other places we're using `Vec::push`. I dug into the generated code to see what's going on. In the tight loop a big win is likely not having to jump out to `RawVec::grow_one`. Nice find! <img alt="Screenshot 2025-04-18 at 10 32 46 AM" width="1131" src="https://private-user-images.githubusercontent.com/6855576/435219477-db0391da-717f-4d06-994e-26fe21358aba.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NDQ5ODcyMzQsIm5iZiI6MTc0NDk4NjkzNCwicGF0aCI6Ii82ODU1NTc2LzQzNTIxOTQ3Ny1kYjAzOTFkYS03MTdmLTRkMDYtOTk0ZS0yNmZlMjEzNThhYmEucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI1MDQxOCUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNTA0MThUMTQzNTM0WiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9MjI4ZDcyMWRmNjI1MDIzYmEwMDcxZDdmNDFjMjFjM2M4NTlhOGVkYWM4ZWJhZGMxZmVhYzE3NjQ1OTAzNTUzNSZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QifQ.-2aVPfXMDDWEhrTZ7GzPVyDUgAskZ6QVkVk-4-2uPkw"> [pr.txt](https://github.com/user-attachments/files/19813348/pr.txt) [main.txt](https://github.com/user-attachments/files/19813349/main.txt)

Thanks for checking! I wonder, is this screenshot from Compiler Explorer or a local tool?