@Dandandan 2025-04-18
> > Thanks for checking! I wonder, is this screenshot from Compiler Explorer or a local tool?
> 
> It's [Beyond Compare](https://www.scootersoftware.com) but you can use your diff tool of choice with the two .txt dumps of the machine code. I use [cargo-show-asm](https://github.com/pacak/cargo-show-asm) to generate the relevant snippets of code.
> 
> For example, `cargo asm -p arrow-select --lib --simplify` shows all the functions, and then I look for `interleave_primitive`. It was function 934 so I did `cargo asm -p arrow-select --lib --simplify 934 > ~/Desktop/pr.txt` to get the pr.txt file. You may need to add the `--context` argument if your function calls out to other functions, or look for functions that call your function if it's been inlined.
> 
> I don't find that Compiler Explorer generates representative code for real projects. Putting snippets in there often doesn't reflect what the compiler does with large projects with complex CFG DAGs, external crates, LTO, and inlined functions.

Ah nice, thanks for the overview! I'll try might try that in the future.