@Dandandan 2025-04-21
> > For example, `cargo asm -p arrow-select --lib --simplify` shows all the functions, and then I look for `interleave_primitive`. It was function 934 so I did `cargo asm -p arrow-select --lib --simplify 934`
> 
> If you know the name of the function or even part of it - you can specify all of it or parts as well: `cargo asm -p arrow-select --lib --simplify interleave_pr`, it should print only names that contain this substring - it should be easier to find the name you want since there will be less output. And then you can pass the number along with the name: `cargo asm -p arrow-select --lib --simplify interleave_pr 934`

TY @pacak !